Put the maven assembly stanza back in.
I think that by running:

mvn assembly:assembly

we now generate essentially the jar that we want to drop into Fedora.
It may be that we need to tweak what gets into this (say by
changing some of the dependency scopes to 'provided') to account
for the jars that fedora is already providing, but I think we're
getting to where we want to be.