After a CaringoBlob does an Scsp action it holds on to the response (replacing a previous one). When we do a write action we use the CaringoHints to generate custom metadata headers. At present this is very simple - the keys of the hints are of the form 'namespace:key' which gets mapped to x-namespace-meta-key for the header name. This may not support everything we want, but it's a way to get started here. Added test to make sure that these headers get added. They do.
Future questions:
- How to support repeated headers (e.g. for Lifepoints we may need this).
An idea would be to allow the values for CaringoHint to be strings that
have a separator character to allow each to hold multiple values. Then
add one for each value that is held. Of course some sort of reasonable
quoting mechanism would be needed.
- How to support other headers besides x-*-meta-*. Maybe augment the
key syntax: namespace:key works as before, but :name maps directly to
the name. In fact, this part should be easy to support, so I think I'll
do it presently.