Add classes that will do Fedora specific behaviors and a test. Abstracted some common test functionality - there may be more to do here.
As things are now, the FedoraBlobStore takes an extra constructor parameter corresponding to the repository name. It then adds that as a hint.
FedoraBlob adds two hints - one to say that this is a fedora stream (redundant given the other hints - I may remove it) and one
to hold the fedora stream id.

I think we may be able to tell if the stream represents a proper fedora object or a managed stream from the id - if so I may put in a separate marker header for this. So FedoraBlob would use the id to determine this and insert the appropriate header. Still need to experiment a bit on this, though.