Move all hint stuff from the Caringo-* hierarchy to Hinted-*, and have Fedora-* inherit from those.
Refactor the tests accordingly.

Now the Caringo-* stuff will work as a simple Akubra adapter without trying to do anything with hints, as it did before. The Hinted-* stuff does our custom hint stuff, and Fedora-* uses it to do some special Fedora stuff.

Note that I shadow an instance variable in HintedBlob (in order to change its type). I think this might be a bit dodgy, but I'm not sure what the Java way to do this is - in something like Smalltalk or Ruby it'd be a non-issue.