I think I have things sorted now - the blob id is exactly the x-fedora-meta-stream-id header, not with info:fedora/ removed from the beginning. Once this is fixed the existence check works.
Note that I'm note sure how the repo rebuilder wants to see the ids
that it gets. To be consistent it _should_ want to see the full ones
that it made the blobs with, but who knows. If there is a problem then
we'll need to fix up the returned ids to conform to expectations. But
that can wait.